Revert "Move athena webscraper from Endpoint List webscrapers to chpl endpoint parsers since the link for it is contained in the CHPL endpoints list"
This reverts commit 686b839647119d6462705f56cc56edb72358f2f7.
