+++ endpointmanager/pkg/endpointwebscraper/endpointwebscraper.go
@@ -12,6 +12,7 @@ type LanternEntry struct {
 
 var oneUpURL = "https://1up.health/fhir-endpoint-directory"
 var careEvolutionURL = "https://fhir.docs.careevolution.com/overview/public_endpoints.html"
+var athenaHealthURL = "https://mydata.athenahealth.com/aserver"
 var techCareURL = "https://devportal.techcareehr.com/Serviceurls"
 var carefluenceURL = "https://carefluence.com/carefluence-fhir-endpoints/"
 
@@ -19,6 +20,8 @@ func EndpointListWebscraper(vendorURL string, vendor string, fileToWriteTo strin
 
 	if vendorURL == oneUpURL || vendorURL == careEvolutionURL {
 		HTMLtablewebscraper(vendorURL, vendor, fileToWriteTo)
+	} else if vendorURL == athenaHealthURL {
+		Athenawebscraper(vendorURL, fileToWriteTo)
 	} else if vendorURL == techCareURL {
 		Techcarewebscraper(vendorURL, fileToWriteTo)
 	} else if vendorURL == carefluenceURL {
