+++ resources/dev_resources/EndpointResourcesList.json
@@ -25,6 +25,12 @@
     },
     {
         "FormatType": "Lantern",
+        "EndpointName": "AthenaHealth",
+        "FileName": "AthenaHealthEndpointSources.json",
+        "URL": "https://mydata.athenahealth.com/aserver"
+    },
+    {
+        "FormatType": "Lantern",
         "EndpointName": "TechCare",
         "FileName": "TechCareEndpointSources.json",
         "URL": "https://devportal.techcareehr.com/Serviceurls"
